Fix OOM in invalidation job when tile coverage is large
After I removed rmdir for `__mosaic__` directory in f991aec7a4fefcb03a2f146fc651f6909af52117
server started to fail because of OOM caused by cache invalidation job. The reason for that
is that if there are millions tiles to invalidate there will be millions of promise objects
allocated for rm calls. This might be solved by deleting files sequentially or limiting
concurrency but still it will take too much to handle millions of rm calls.

Also I figured that removing whole mosaic cache directory is not an option because it will
take too much time to rebuild.

Here I want to try another approach: if tile coverage of area that need to be invalidated is
too large there is high probablity that most of these millions of tile in coverage do not
even present in the cache so it might make sense to go through the list of items in the cache
and check if they need to be invalidated rather then trying to delete all items from invalidated
area tile coverage.